+++ src/infi/clickhouse_orm/models.py
@@ -311,9 +311,16 @@ class MergeModel(Model):
 
     @classmethod
     def create_table_sql(cls, db):
-        assert isinstance(cls.engine, Merge), "engine must be engines.Merge instance"
-        return super(MergeModel, cls).create_table_sql(db)
-
+        assert isinstance(cls.engine, Merge), "engine must be an instance of engines.Merge"
+        parts = ['CREATE TABLE IF NOT EXISTS `%s`.`%s` (' % (db.db_name, cls.table_name())]
+        cols = []
+        for name, field in iteritems(cls.fields()):
+            if name != '_table':
+                cols.append('    %s %s' % (name, field.get_sql()))
+        parts.append(',\n'.join(cols))
+        parts.append(')')
+        parts.append('ENGINE = ' + cls.engine.create_table_sql(db))
+        return '\n'.join(parts)
 
 # TODO: base class for models that require specific engine
 
@@ -324,7 +331,7 @@ class DistributedModel(Model):
     """
 
     def set_database(self, db):
-        assert isinstance(self.engine, Distributed), "engine must be engines.Distributed instance"
+        assert isinstance(self.engine, Distributed), "engine must be an instance of engines.Distributed"
         res = super(DistributedModel, self).set_database(db)
         return res