+++ tests/test_custom_fields.py
@@ -9,7 +9,7 @@ from infi.clickhouse_orm.engines import Memory
 class CustomFieldsTest(unittest.TestCase):
 
     def setUp(self):
-        self.database = Database('test-db')
+        self.database = Database('test-db', log_statements=True)
 
     def tearDown(self):
         self.database.drop_database()